{"schema":"decodex/commit/1","summary":"Move window monitor enumeration","authority":"manual"}
